+++ tests/aiobungie/test_client.py
@@ -39,8 +39,8 @@ _LOG = logging.getLogger("test_client")
 
 def __build_client() -> aiobungie.Client:
     token = os.environ["CLIENT_TOKEN"]
-    rest = aiobungie.RESTClient(token, max_retries=0, enable_debugging=True)
-    client = aiobungie.Client(token, rest_client=rest, max_retries=0)
+    client = aiobungie.Client(token, max_retries=0)
+    client.rest.enable_debugging(True)
     return client